Refactor libs to align with fbcode build
