+++ lib/Optimizer/CMakeLists.txt
@@ -1,3 +1,4 @@
 add_subdirectory(GraphOptimizer)
+add_subdirectory(GraphOptimizerPipeline)
 
 add_subdirectory(IROptimizer)
